added float cast for trying to build release.

+++ vector3d.cpp
@@ -96,7 +96,7 @@ float* vector3d::to_sph(){
 			return this->abc; //it's a null vector, return a null vector.
 		}
 		float t = atan2(this->abc[1] , this->abc[0]);
-		t = (t < 0) ? (t + (2 * M_PI)) : t; //if the result is negative, then push into 0-2PI range
+		t = (t < 0) ? (float)(t + (2 * M_PI)) : t; //if the result is negative, then push into 0-2PI range
 		float p = acos(this->abc[2] / r);
 		this->abc[0] = r;
 		this->abc[1] = t;
